requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.8
django-allauth==65.13.1
idna==3.11
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.4
urllib3==2.5.0
